feat: 무한 스크롤 추가 및 모임 목록 컴포넌트 변경
